Refined Whiteboarding Code.
